data transfer
When the user clicks on the table cell, the didSelectRowAt function is activated. This function retrieves the name and ID of the user's selected item from the nameArray and idArray arrays and then switches to the details view by calling the performSegue function. The transition to the details view is prepared in the prepare function. This function checks whether the correct transition has occurred by checking the identity of the segue object. If the transition has been made correctly, the target view is retrieved with segue.destination and converted to DetailsVC type. Then, the name and ID of the selected image are passed to this destination view. Additionally, an observer (NotificationCenter.default.addObserver(...)) is added to call the getData function when new data is received. This ensures that new data is automatically retrieved if the data is updated.